Sounds, Script Editor, Alpha masking SIKE!
... Well, kind of...

There are now `Sound` objects, but they don't work and idk why

The Editor can now let you edit scripts, but like why would you use it??

Also fixed a memory leak in `FileRW`, and some slight cleanup in the Editor

Oh and `worldUber` supports a new "AlphaCutoff" uniform, which is set in the Model Loader for materials which use the "MASK" transparency type.

+++ src/impl/asset/MeshProvider.cpp
@@ -371,5 +371,5 @@ void MeshProvider::m_CreateAndUploadGpuMesh(Mesh& mesh)
 	mesh.Vertices.clear();
 	mesh.Indices.clear();
 
-	mesh.GpuId = m_GpuMeshes.size() - 1;
+	mesh.GpuId = static_cast<uint32_t>(m_GpuMeshes.size() - 1);
 }
